:recycle: refactor: restructure markdown component
