+++ frontend/src/components/index.ts
@@ -1,4 +1,5 @@
 export { Breadcrumbs } from "./Breadcrumbs";
 export { LabeledIcon } from "./LabeledIcon";
+export { Markdown } from "./Markdown";
 export { Logo } from "./Logo";
 export { default as Link, type LinkProps, NextLinkComposed } from "./Link";
